don't call it a hack
the joking nature of my language was a bit mistaken. I don't consider this a code a single ugly hack. A hack is if I differ between a 32 bit pointer and a 31 bit value by setting the LSB to 1 ;-) or maybe using a simpler approach over ERM (which i want to rectify).